remove zero on-hand inventory records

+++ src/main/java/guru/sfg/beer/inventory/service/services/AllocationServiceImpl.java
@@ -77,7 +77,9 @@ public class AllocationServiceImpl implements AllocationService {
 
                 beerOrderLineDto.setQuantityAllocated(orderLineAllocated + inventory);
                 inventoryItem.setQuantityOnHand(0);
+            }
 
+            if (inventoryItem.getQuantityOnHand() == 0) {
                 beerInventoryRepository.delete(inventoryItem);
             }
         });
